# batch_norm_cpu_inference for channel last