--- a/aten/src/ATen/native/Normalization.cpp
+++ b/aten/src/ATen/native/Normalization.cpp
@@ -56,20 +56,12 @@ struct Var {
   }
 };
 
-/// A fast path for CPU inference when all tensors are contiguous.
-/// This code achieves machine bandwidth peak without AVX support.
-/// If this changes for future architectures, we can move it to the cpu/
-/// directory.
 template<typename scalar_t>
-void batch_norm_cpu_inference_contiguous(Tensor& output, const Tensor& input,
+void batch_normal_cpu_inference_collect_liner_and_constant_terms(
+    scalar_t* alpha, scalar_t* beta, int64_t n_channel,
     const Tensor& weight /* optional */, const Tensor& bias /* optional */,
     const Tensor& mean, const Tensor& variance, double eps) {
-  int64_t n_batch = input.size(0);
-  int64_t n_channel = input.size(1);
-  int64_t image_size = input.numel() / n_batch / n_channel;
 
-  scalar_t* output_data = output.data_ptr<scalar_t>();
-  const scalar_t* input_data = input.data_ptr<scalar_t>();
   const scalar_t* weight_data = weight.defined() ? weight.data_ptr<scalar_t>() : nullptr;
   const scalar_t* bias_data = bias.defined() ? bias.data_ptr<scalar_t>() : nullptr;
   const scalar_t* mean_data = mean.data_ptr<scalar_t>();
@@ -86,17 +78,38 @@ void batch_norm_cpu_inference_contiguous(Tensor& output, const Tensor& input,
   ///   the constant term beta(c) = bias(c) - mean(c) * inv_var(c) * weight(c)
   /// Note that this is only a good idea if (input_size >> c), in degenerate
   /// cases where image_size == 1 && batch_size == 1, it is slow.
-  Tensor alpha = at::empty_like(mean);
-  Tensor beta = at::empty_like(mean);
-  scalar_t* alpha_data = alpha.data_ptr<scalar_t>();
-  scalar_t* beta_data = beta.data_ptr<scalar_t>();
   for (int64_t c = 0; c < n_channel; c++) {
     scalar_t inv_var = 1 / std::sqrt(var_data[c] + static_cast<scalar_t>(eps));
     scalar_t weight_v = weight_data ? weight_data[c] : 1;
     scalar_t bias_v = bias_data ? bias_data[c] : 0;
-    alpha_data[c] = inv_var * weight_v;
-    beta_data[c] = bias_v - mean_data[c] * inv_var * weight_v;
+    alpha[c] = inv_var * weight_v;
+    beta[c] = bias_v - mean_data[c] * inv_var * weight_v;
   }
+}
+
+/// A fast path for CPU inference when all tensors are contiguous.
+/// This code achieves machine bandwidth peak without AVX support.
+/// If this changes for future architectures, we can move it to the cpu/
+/// directory.
+template<typename scalar_t>
+void batch_norm_cpu_inference_contiguous(Tensor& output, const Tensor& input,
+    const Tensor& weight /* optional */, const Tensor& bias /* optional */,
+    const Tensor& mean, const Tensor& variance, double eps) {
+
+  int64_t n_batch = input.size(0);
+  int64_t n_channel = input.size(1);
+  int64_t image_size = input.numel() / n_batch / n_channel;
+
+  scalar_t* output_data = output.data_ptr<scalar_t>();
+  const scalar_t* input_data = input.data_ptr<scalar_t>();
+
+  Tensor alpha = at::empty_like(mean);
+  Tensor beta = at::empty_like(mean);
+  scalar_t* alpha_data = alpha.data_ptr<scalar_t>();
+  scalar_t* beta_data = beta.data_ptr<scalar_t>();
+
+  batch_normal_cpu_inference_collect_liner_and_constant_terms<scalar_t>(
+      alpha_data, beta_data, n_channel, weight, bias, mean, variance, eps);
 
   // Apply the linear terms to the input,
   // output(n, c, h, w) = input(n, c, h, w) * alpha(c) + beta(c)
@@ -127,6 +140,58 @@ void batch_norm_cpu_inference_contiguous(Tensor& output, const Tensor& input,
   }
 }
 
+/// A fast path for CPU inference when all tensors are channels last contiguous.
+/// This code achieves machine bandwidth peak without AVX support.
+/// If this changes for future architectures, we can move it to the cpu/
+/// directory.
+template<typename scalar_t>
+void batch_norm_cpu_inference_channels_last(Tensor& output, const Tensor& input,
+    const Tensor& weight /* optional */, const Tensor& bias /* optional */,
+    const Tensor& mean, const Tensor& variance, double eps) {
+
+  int64_t n_batch = input.size(0);
+  int64_t n_channel = input.size(1);
+  int64_t image_size = input.numel() / n_batch / n_channel;
+
+  scalar_t* output_data = output.data_ptr<scalar_t>();
+  const scalar_t* input_data = input.data_ptr<scalar_t>();
+
+  Tensor alpha = at::empty_like(mean);
+  Tensor beta = at::empty_like(mean);
+  scalar_t* alpha_data = alpha.data_ptr<scalar_t>();
+  scalar_t* beta_data = beta.data_ptr<scalar_t>();
+
+  batch_normal_cpu_inference_collect_liner_and_constant_terms<scalar_t>(
+      alpha_data, beta_data, n_channel, weight, bias, mean, variance, eps);
+
+  // Apply the linear terms to the input,
+  // output(n, c, h, w) = input(n, c, h, w) * alpha(c) + beta(c)
+  // No need to use parallel_for as this function is supposed to be
+  // memory-limited.
+  // Keep the loop struture simple to make sure compiler vetorization kicks in.
+  if (n_channel != 1) {
+    for (int64_t n = 0; n < n_batch; ++n) {
+      for (int64_t i = 0; i < image_size; ++i) {
+      for (int64_t c = 0; c < n_channel; ++c) {
+          // Keep all the offset calculation within the inner loop for
+          // simplicity. Compilers are very good at hoisting the common part
+          // outside.
+          int64_t offset = n * image_size * n_channel + i * n_channel + c;
+          output_data[offset] = input_data[offset] * alpha_data[c] + beta_data[c];
+        }
+      }
+    }
+  } else {
+    // n_channel == 1
+    for (int64_t n = 0; n < n_batch; ++n) {
+      for (int64_t i = 0; i < image_size; ++i) {
+        int64_t offset = n * image_size + i;
+        output_data[offset] = input_data[offset] * alpha_data[0] + beta_data[0];
+      }
+    }
+  }
+}
+
 template<typename scalar_t>
 std::tuple<Tensor,Tensor,Tensor> batch_norm_cpu_transform_input_template(
     const Tensor& input, const Tensor& weight, const Tensor& bias,
@@ -134,18 +199,34 @@ std::tuple<Tensor,Tensor,Tensor> batch_norm_cpu_transform_input_template(
     const Tensor& running_mean /* optional */, const Tensor& running_var /* optional */,
     bool train, double eps) {
 
-  Tensor output = at::empty_like(input);
-
-  // Check if we should use the fast path.
+  // Check if we should use the fast path for contiguous memory format
   if (!train && input.is_contiguous()
       && (!weight.defined() || weight.is_contiguous())
       && (!bias.defined() || bias.is_contiguous())
       && running_mean.is_contiguous()
       && running_var.is_contiguous()) {
-    batch_norm_cpu_inference_contiguous<scalar_t>(output, input, weight, bias,
-      running_mean, running_var, eps);
+
+    Tensor output = at::empty_like(input);
+    batch_norm_cpu_inference_contiguous<scalar_t>(
+      output, input, weight, bias, running_mean, running_var, eps);
     return std::make_tuple(output, save_mean, save_invstd);
   }
+
+  // Check if we should use the fast path for channel last memory format
+  if (!train && input.is_contiguous(at::MemoryFormat::ChannelsLast)
+      && (!weight.defined() || weight.is_contiguous())
+      && (!bias.defined() || bias.is_contiguous())
+      && running_mean.is_contiguous()
+      && running_var.is_contiguous()) {
+
+    Tensor output = at::empty_like(input, at::MemoryFormat::ChannelsLast);
+    batch_norm_cpu_inference_channels_last<scalar_t>(
+      output, input, weight, bias, running_mean, running_var, eps);
+    return std::make_tuple(output, save_mean, save_invstd);
+  }
+
+  Tensor output = at::empty_like(input);
+
   int64_t n_input = input.size(1);
 
   auto save_mean_a = conditional_accessor_1d<scalar_t>(save_mean);