--- a/aten/src/ATen/native/Normalization.cpp
+++ b/aten/src/ATen/native/Normalization.cpp
@@ -57,7 +57,7 @@ struct Var {
 };
 
 template<typename scalar_t>
-void batch_normal_cpu_inference_collect_liner_and_constant_terms(
+void batch_norm_cpu_inference_collect_liner_and_constant_terms(
     scalar_t* alpha, scalar_t* beta, int64_t n_channel,
     const Tensor& weight /* optional */, const Tensor& bias /* optional */,
     const Tensor& mean, const Tensor& variance, double eps) {
@@ -108,7 +108,7 @@ void batch_norm_cpu_inference_contiguous(Tensor& output, const Tensor& input,
   scalar_t* alpha_data = alpha.data_ptr<scalar_t>();
   scalar_t* beta_data = beta.data_ptr<scalar_t>();
 
-  batch_normal_cpu_inference_collect_liner_and_constant_terms<scalar_t>(
+  batch_norm_cpu_inference_collect_liner_and_constant_terms<scalar_t>(
       alpha_data, beta_data, n_channel, weight, bias, mean, variance, eps);
 
   // Apply the linear terms to the input,
@@ -161,7 +161,7 @@ void batch_norm_cpu_inference_channels_last(Tensor& output, const Tensor& input,
   scalar_t* alpha_data = alpha.data_ptr<scalar_t>();
   scalar_t* beta_data = beta.data_ptr<scalar_t>();
 
-  batch_normal_cpu_inference_collect_liner_and_constant_terms<scalar_t>(
+  batch_norm_cpu_inference_collect_liner_and_constant_terms<scalar_t>(
       alpha_data, beta_data, n_channel, weight, bias, mean, variance, eps);
 
   // Apply the linear terms to the input,